address comments: rename, breakout tests
